Introduction

Welcome to the documentation of Lattice.jl

This is a Julia package for defining lattices in Julia.

All lattice type in this package is duck typed. Any Julia type with the (Geometric) Region/Lattice Interface can be considered as a (geometric) region/lattice.

Interfaces

Site iterator

function: sites

Edge iterator

function: edges

Surround iterator (Optional)

function: surround

Face iterator (Optional)

function: faces

Lattices

Lattices.BoundedLattice — Type.

BoundedLattice{N, BC} <: AbstractLattice{N}

A lattice with boundary conditions BC on N dimension.

source

Lattices.Chain — Type.

Chain{L, BC} <: AbstractLattice

Chain lattice. L is the length of this chain, and BC is the boundary condition.

source

Lattices.Chain — Method.

Chain(length; [boundary=Periodic])

Returns a chain lattice with given length, the boundary is set to Periodic by default.

source

Lattices.Fixed — Type.

Fixed <: BoundaryCondition

Fixed boundary condition.

source

Lattices.Periodic — Type.

Periodic <: BoundaryCondition

Periodic boundary condition.

source

Lattices.Square — Type.

Square{Shape, BC} <: BoundedLattice{2, BC}

Square lattice.

source

Lattices.Square — Method.

Square(height, width; [boundary=Periodic])

Returns a square lattice with given height and width, boundary is set to Periodic by default.

source

Lattices.boundary — Function.

boundary(lattice) -> BoundaryCondition

Returns lattice's boundary condition.

source

Lattices.edges — Function.

edges(lattice; [length=1]) -> iterator

Returns an iterator for all edges of length on the lattice.

source

Lattices.isperiodic — Function.

isperiodic(lattice) -> Bool

Check if this lattice's boundary condition is periodic.

source

Lattices.sites — Function.

sites(lattice) -> iterator

Returns an iterator for all sites on the lattice.

source

Lattices.AbstractLattice — Type.

AbstractLattice{N}

Abstract type for general lattices. N indicates the dimension. For a more concrete definition please refer the following material:

Lattice (group): https://en.wikipedia.org/wiki/Lattice_(group)

Lattice Model: https://en.wikipedia.org/wiki/Latticemodel(physics)

Lattice Graph: https://en.wikipedia.org/wiki/Lattice_graph

source

Lattices.BoundaryCondition — Type.

BoundaryCondition

Boundary conditions for lattices. There usually has two:

- Fixed boundary
- Periodic boundary

Ref: https://nanohub.org/resources/7577/download/MartiniL5BoundaryConditions.pdf

source

Lattices.SquareEdgesIterator — Type.

SquareEdgesIterator{Region, Order, LT}

iterate through order-th edge in Region on square lattide LT.

source

Base.length — Function.

length(lattice) -> Int

Returns the length of this lattice, or the production of each dimension size.

source

Base.nameof — Method.

name(lattice) -> String

Returns the name of this lattice.

source

Base.ndims — Method.

ndims(lattice) -> N

Returns number of dimensions of the lattice.

source

Base.size — Function.

size(lattice) -> Tuple

Returns the size of this lattice.

source

Lattices.fastmod1 — Method.

fastmod1(::Val{X}, ::Val{Y})

mod1 in compile time, 40x faster if the value can be determined

source